add a new provider type, "WINDOW"
when this provider type is set in the Maker constructor's options, it will look for Metamask, using window.web3 if present and using the new "request provider with postMessage" method otherwise.

This will set up all accounts to send their transactions through Metamask regardless of what account type they are.

the usePresetProvider and overrideMetamask settings are now ignored.