Get testnet DSProxy address from addresses.json
getContractAddressByName() isn’t being used here because a DSProxy by nature doesn’t live at a fixed address, but separate addresses for each owner’s instance, so it wouldn’t make sense to have it included in the address mappings. It’s being deployed in the deployment script more as a type of test than anything.