ensure QueryApi.getCdpIdsForOwner converts the address arg to a checksumAddress
Summary:

the dai.js library code seems to convert most addresses to lowercase. This is
presumably done to make it canonical, and perhaps easier to compare. Not entirely
sure.

However, when querying the server via graphql, the server expects a checksumAddress
with the appropriate characters uppercase and lowercase. See EIP-55 for details.
https://github.com/ethereum/EIPs/blob/master/EIPS/eip-55.md

So, this change does that.

Ideally, we'd refactor the code so that it always uses this checksum-encoding
for the address, but that seems like a larger change than i am comfortable
making at the moment.

Test Plan:

this code works for me when i call it in my own repo using the @makerdao/dai
library. Um, i basically copy-pasta'd the entire QueryApi code into my repo,
and modified it as shown here.

Prior to having this code, any call to `getCdpIdsForOwner` would return an empty array.